fix the displaying of the error dialogs
Exceptions which get raised in threads don't bubble up into the main code on
their own. So we either communicate them to the main loop or we handle them
directly in the thread. I'm going for the later for the moment because the
other way seems to be quite complex.