reimplement the run method on the backup class
For some reason the BorgExceptions from the backup thread don't get caught in
the main_window. Therefore I'm implementing the error handling directly in the
class for the moment. This will have the side effect that the error dialog will
pop up when there is a problem during a background backup.

However that is anyway something I thought about implementing. This way it's
just not properly implemented yet.